Рефакторинг movie-genres-list
