+++ src/components/movie-genres-list/movie-genres-list.jsx
@@ -1,36 +1,22 @@
 import React from "react";
 import PropTypes from "prop-types";
+import MovieGenresItem from "../movie-genres-item/movie-genres-item";
 
 
 const MovieGenresList = (props) => {
-
-  const {genresList, activeGenre, onGenreClick} = props;
-
-  const onButtonClick = (evt) => {
-    evt.preventDefault();
-    const genre = evt.target.textContent;
-    onGenreClick(genre);
-  };
+  const {genresList} = props;
 
   return (
     <ul className="catalog__genres-list">
-
       {genresList.map((genre, i) => (
-        <li key ={i + genre} className={`catalog__genres-item ${(genre === activeGenre) ? `catalog__genres-item--active` : ``}`}>
-          <a href="#" className="catalog__genres-link" onClick={onButtonClick}>
-            {genre}
-          </a>
-        </li>
+        <MovieGenresItem key ={i + genre} {...props} genre={genre} />
       ))}
-
     </ul>
   );
 };
 
 MovieGenresList.propTypes = {
   genresList: PropTypes.arrayOf(PropTypes.string).isRequired,
-  activeGenre: PropTypes.string.isRequired,
-  onGenreClick: PropTypes.func.isRequired,
 };
 
 export default MovieGenresList;
